refactor: split database and query sql
